% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{find.pred.stats}
\alias{find.pred.stats}
\title{find.pred.stats}
\usage{
find.pred.stats(data, yvar, trtvar, type, censorvar)
}
\arguments{
\item{data}{data frame with response and prediction vector}

\item{yvar}{response variable name}

\item{trtvar}{treatment variable name}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c".}

\item{censorvar}{censoring variable name}
}
\value{
a data frame of predictive statistics
}
\description{
Find predictive stats from response and prediction vector
}
\details{
Find predictive stats from response and prediction vector
}
