% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{permute.rows}
\alias{permute.rows}
\title{permute.rows}
\usage{
permute.rows(A)
}
\arguments{
\item{A}{a matrix for which its rows have to be permuted.}
}
\value{
the matrix with permuted rows.
}
\description{
Randomly permute the rows of a matrix.
}
\details{
Randomly permute the rows of a matrix.
}
